extracts 'show' functionality to cumulative code execution hook
