+++ src/components/code-cell/code-cell.tsx
@@ -2,7 +2,7 @@ import './code-cell.css'
 import { useEffect } from 'react'
 import { CodeEditor, Preview, Resizable } from '..'
 import { Cell } from '../../state'
-import { useActions, useTypedSelector } from '../../hooks'
+import { useActions, useCumulativeCode, useTypedSelector } from '../../hooks'
 
 interface CodeCellProps {
   cell: Cell
@@ -11,66 +11,25 @@ interface CodeCellProps {
 const CodeCell: React.FC<CodeCellProps> = ({ cell }) => {
   const { updateCell, createBundle } = useActions()
   const bundle = useTypedSelector(state => state.bundles[cell.id])
-
-  const cumulativeCode = useTypedSelector(state => {
-    const { data, order } = state.cells
-    const orderedCells = order.map(id => data[id])
-
-    const showFn = `
-    import _React from 'react'
-    import _ReactDOM from 'react-dom'
-
-    var show = (value) => {
-      const root = document.querySelector('#root')
-      
-      if (typeof  value === 'object') {
-        if (value.$$typeof && value.props) {
-          _ReactDOM.render(value, root)
-        } else {
-          root.innerHTML = JSON.stringify(value)
-        }
-      } else {
-        root.innerHTML = value
-      }
-    }
-  `
-
-    const showFnNoOp = 'var show = () => {}'
-
-    const cumulativeCode = []
-    for (let c of orderedCells) {
-      if (c.type === 'code') {
-        if (c.id === cell.id) {
-          cumulativeCode.push(showFn)
-        } else {
-          cumulativeCode.push(showFnNoOp)
-        }
-        cumulativeCode.push(c.content)
-      }
-      if (c.id === cell.id) {
-        break
-      }
-    }
-    return cumulativeCode
-  })
+  const cumulativeCode = useCumulativeCode(cell.id)
 
   // auto-bundle user code after 0.8s
   // if there is no bundle, immediately create it
   useEffect(() => {
     if (!bundle) {
-      createBundle(cell.id, cumulativeCode.join('\n'))
+      createBundle(cell.id, cumulativeCode)
       return
     }
 
     const timer = setTimeout(async () => {
-      createBundle(cell.id, cumulativeCode.join('\n'))
+      createBundle(cell.id, cumulativeCode)
     }, 800)
 
     return () => {
       clearTimeout(timer)
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [cumulativeCode.join('\n'), cell.id, createBundle])
+  }, [cumulativeCode, cell.id, createBundle])
 
   return (
     <Resizable direction='vertical'>
