+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './use-actions'
 export * from './use-typed-selector'
+export * from './use-cumulative-code'
